+++ src/runtime/codemirror-language.js
@@ -1,10 +1,21 @@
 const language = acode.require("@codemirror/language");
 
 export const {
+	Language,
 	LanguageSupport,
 	LRLanguage,
+	bracketMatching,
+	bracketMatchingHandle,
+	continuedIndent,
+	defineLanguageFacet,
 	foldInside,
 	foldNodeProp,
+	flatIndent,
+	getIndentation,
+	getIndentUnit,
+	indentString,
 	indentNodeProp,
 	delimitedIndent,
+	sublanguageProp,
+	syntaxTree,
 } = language;